prj2

+++ proj2/byog/Core/Game.java
@@ -43,7 +43,7 @@ public class Game {
         // TODO: Fill out this method to run the game using the input passed in,
         //      and return a 2D tile representation of the world that would have been
         //      drawn if the same inputs had been given to playWithKeyboard().
-        int seed = Integer.parseInt(input.substring(1,input.length()-1));
+        long seed = Long.parseLong(input.substring(1,input.length()-1));
         RANDOM = new Random(seed);
         TETile[][] finalWorldFrame = null;
         switch (input.charAt(0)) {
